chore: rename files and address pr comments
