+++ packages/backend/src/apps/webhook/actions/make-request/index.ts
@@ -7,7 +7,7 @@ import { isUrlAllowed } from '../../common/ip-resolver'
 type TMethod = 'GET' | 'POST' | 'PATCH' | 'PUT' | 'DELETE'
 
 export default defineAction({
-  name: 'HTTP Request',
+  name: 'HTTP Request (Deprecated: please use the one under Custom API app)',
   key: 'httpRequest',
   description: 'Makes a custom HTTP request by providing raw details.',
   arguments: [
